Fix pinMode() in Arduino HAL
The Arduino pinMode(pin, mode) functions the specified pin
into a LOW state if the wanted mode is OUTPUT.

I wrongly assumed that I can put the pin HIGH and that the code
will use a digitalWrite() to set the desired value.

This was wrong, as e.g. seen in TM1637.cpp:191.

With this fix, the module now displays stuff correctly!

Also decreases bit delay times since it works fine with the fast
original speed.